Move AbstractNotifier-tests to MailNotifierTests
Tests in Abstract*.class are not executed during the maven build, so I moved
the tests for code in AbstractNotifierTest to MailNotifierTest
